bersih bersih

+++ inventory.php
@@ -77,11 +77,20 @@
 ?>
 <?php
     if (isset($_POST['delID'])) {
-        $sql_delBrg = "DELETE FROM item WHERE Id = :ni";
-        $stmt_delBrg = $conn->prepare($sql_delBrg);
-        $stmt_delBrg->execute(array(
+        $sql_cek = "SELECT STATUS FROM item WHERE Id = :ni";
+        $stmt_cek = $conn->prepare($sql_cek);
+        $stmt_cek->execute(array(
             ":ni" => $_POST['delID']
         ));
+        $row = $stmt_cek->fetchColumn();
+        if($row == 1){
+            $sql_delBrg = "DELETE FROM item WHERE Id = :ni";
+            $stmt_delBrg = $conn->prepare($sql_delBrg);
+            $stmt_delBrg->execute(array(
+                ":ni" => $_POST['delID']
+            ));
+        }
+        echo $row;
         exit();
     }
 ?>
@@ -293,10 +302,6 @@
                         aksi : 1
                     },
                     success : function(e) {
-                        console.log(kode);
-                        console.log(kode_bor);
-                        console.log(`${result.value.user}`);
-                        console.log(e);
                         if (e == 1) {
                             swalWithBootstrapButtons.fire ({
                                     icon : "success",
@@ -396,7 +401,6 @@
                         newDesc : $("#newDesc").val()
                     },
                     success : function(e) {
-                        console.log();
                         showResult(e);
                         $("#table").DataTable().ajax.reload(null,false);
                         $("#exampleModal").modal('toggle');
@@ -422,21 +426,36 @@
                     confirmButtonText: 'Continue',
                     cancelButtonText: 'Cancel',
                 }).then((result) => {
-                    $.ajax({
-                    type : "post",
-                    data : {
-                        delID : id
-                    },
-                    success : function(e){
-                        console.log(e);
+                    if(result.isConfirmed){
+                        $.ajax({
+                            type : "post",
+                            data : {
+                                delID : id
+                            },
+                            success : function(e){
+                                if(e == 1){
+                                    swalWithBootstrapButtons.fire({
+                                        icon : "success",
+                                        title : "Success!",
+                                        text : "Data Deleted!"
+                                    });
+                                }else{
+                                    swalWithBootstrapButtons.fire({
+                                        icon : "error",
+                                        title : "Failed!",
+                                        text : "Item can't be deleted because there's still an ongoing progress!"
+                                    });
+                                }
+                                $("#table").DataTable().ajax.reload();
+                            }
+                        })
+                    }else{
                         swalWithBootstrapButtons.fire({
-                            icon : "success",
-                            title : "Success!",
-                            text : "Data Deleted!"
+                            icon : "error",
+                            title : "Cancelled!",
+                            text : "Cancel delete item!"
                         });
-                        $("#table").DataTable().ajax.reload();
                     }
-                })
                 })
             });
